JsonSerializableBase: handle objects in `detached` state
The __json__() handles deleted objects, but doesn't handle objects in
the `detached` state. Detached objects may have lazy-loaded attributes,
but such attributes cannot actually be loaded, because a detached object
has no DB Session associated with it.

Prior to SQLAlchemy v1.1 that didn't cause any problems on our projects,
because the old Session.commit() implementation always marked objects
as deleted.

After SQLAlchemy v.1.1, the .commit() method may detach an object from
the session, but not yet mark it as `deleted`. The object becomes
deleted only when the data is deleted from the DB.

So this git commit adds the code that handles object states properly.